+++ Model/Planet.cs
@@ -8,9 +8,9 @@ public readonly record struct Planet
     public string Name { get; }
     public int Diameter { get; }
     public int? SurfaceWater { get; }
-    public int? Population { get; }
+    public long? Population { get; }
 
-    public Planet(string name, int diameter, int? surfaceWater, int? population)
+    public Planet(string name, int diameter, int? surfaceWater, long? population)
     {
         if (name is null)
         {
@@ -28,7 +28,7 @@ public readonly record struct Planet
         var name = planetDto.name;
         var diameter = int.Parse(planetDto.diameter);
 
-        int? population = planetDto.population.ToIntOrNull();
+        long? population = planetDto.population.ToLongOrNull();
         int? surfaceWater = planetDto.surface_water.ToIntOrNull();
 
         return new Planet(name, diameter, surfaceWater, population);
